feature (resource): create an abstract base class that can be extended by resource class specific component classes
