+++ src/app/resource/letter/letter.component.html
@@ -1,5 +1,5 @@
-<kui-progress-indicator *ngIf="loading"></kui-progress-indicator>
-<div class="full-size scroll-y-only" *ngIf="!loading && props">
+<kui-progress-indicator *ngIf="isLoading"></kui-progress-indicator>
+<div class="full-size scroll-y-only" *ngIf="!isLoading && props">
 
     <div class="sidebar">
